+++ polaris/tasks/ocean/external_gravity_wave/init.py
@@ -2,7 +2,7 @@ import numpy as np
 import xarray as xr
 from mpas_tools.io import open_dataset
 
-from polaris.ocean.coriolis import add_coriolis_to_dataset
+from polaris.coriolis import add_coriolis_to_dataset
 from polaris.ocean.model import OceanIOStep
 from polaris.ocean.vertical import init_vertical_coord
