+++ polaris/tasks/ocean/horiz_press_grad/init.py
@@ -4,7 +4,7 @@ from mpas_tools.io import write_netcdf
 from mpas_tools.mesh.conversion import convert, cull
 from mpas_tools.planar_hex import make_planar_hex_mesh
 
-from polaris.ocean.coriolis import add_coriolis_to_dataset
+from polaris.coriolis import add_coriolis_to_dataset
 from polaris.ocean.model import OceanIOStep
 from polaris.ocean.vertical.pstar import init_pstar_vertical_coord
 from polaris.ocean.vertical.pstar_init import PStarInitStep
