+++ polaris/tasks/ocean/inertial_gravity_wave/init.py
@@ -4,8 +4,8 @@ from mpas_tools.io import write_netcdf
 from mpas_tools.mesh.conversion import convert, cull
 from mpas_tools.planar_hex import make_planar_hex_mesh
 
+from polaris.coriolis import add_coriolis_to_dataset
 from polaris.mesh.planar import compute_planar_hex_nx_ny
-from polaris.ocean.coriolis import add_coriolis_to_dataset
 from polaris.ocean.model import OceanIOStep
 from polaris.ocean.surface_pressure import surface_pressure_from_config
 from polaris.ocean.vertical import init_vertical_coord
